feat(client): save and restore snapshots
