+++ src/client/src/rt-platforms/openfin/components/OpenFinSnapshot/ToggleOpenFinLayoutLock.tsx
@@ -1,0 +1,30 @@
+import React, { useState } from 'react'
+import styled from 'styled-components/macro'
+import { toggleLockedLayout } from './utils'
+
+const Container = styled('div')`
+  padding: 0.75rem 0.75rem 0 0.75rem;
+`
+
+const LockButton = styled('button')`
+  &:hover {
+    color: ${({ theme }) => theme.accents.positive.base};
+  }
+`
+
+const ToggleOpenFinLayoutLock = () => {
+  const [locked, setLocked] = useState(false)
+
+  const toggle = () => {
+    setLocked(!locked)
+    toggleLockedLayout()
+  }
+
+  return (
+    <Container>
+      <LockButton onClick={toggle}>{locked ? 'Unlock Layout' : 'Lock Layout'}</LockButton>
+    </Container>
+  )
+}
+
+export default ToggleOpenFinLayoutLock
