+++ src/client/src/rt-platforms/openfin/components/OpenFinSnapshot/utils.tsx
@@ -62,3 +62,54 @@ export const saveSnapshotToStorage = async (newSnapshotName: string) => {
   setSnapshotNames([...snapshotNames, newSnapshotName])
   setSnapshots(snapshots)
 }
+
+/**
+ * Toggles the Platform controls
+ *
+ * Manually created windows are not part of the layout. Layout windows all start
+ * with 'internal-generated', so we find one of those windows to retrieve the
+ * layout.
+ *
+ * https://developers.openfin.co/docs/recipes-platform
+ *
+ */
+export const toggleLockedLayout = async () => {
+  const thisApp = fin.Application.getCurrentSync()
+  const childWindows = await thisApp.getChildWindows()
+  const internallyGeneratedWindows = childWindows.filter(
+    w => w.identity.name && w.identity.name.startsWith('internal-generated')
+  )
+
+  if (internallyGeneratedWindows.length < 1) {
+    console.error('Attempting to toggle layout, but no layout windows found')
+    return
+  }
+
+  const layout = fin.Platform.Layout.wrapSync(internallyGeneratedWindows[0].identity)
+
+  const oldLayout = await layout.getConfig()
+  const { settings, dimensions } = oldLayout
+  if (settings && settings.hasHeaders && settings.reorderEnabled) {
+    layout.replace({
+      ...oldLayout,
+      settings: {
+        ...settings,
+        hasHeaders: false,
+        reorderEnabled: false,
+      },
+    })
+  } else {
+    layout.replace({
+      ...oldLayout,
+      settings: {
+        ...settings,
+        hasHeaders: true,
+        reorderEnabled: true,
+      },
+      dimensions: {
+        ...dimensions,
+        headerHeight: 25,
+      },
+    })
+  }
+}
