+++ src/client/src/rt-platforms/openfin/components/index.ts
@@ -3,4 +3,3 @@ export { OpenFinLogo } from './OpenFinLogo'
 export { default as OpenFinBrowserLink } from './OpenFinBrowserLink'
 export { OpenFinWindowFrame } from './OpenFinWindowFrame'
 export { OpenFinSubWindowFrame } from './OpenFinSubWindowFrame'
-export { default as ToggleOpenFinLayoutLock } from './ToggleOpenFinLayoutLock'
